This fixes the 2 bugs with the older version of Jax (#36 and #58).

I then unpinned the old version of Jax. When I install everything from scratch and run the tests it passes.

Note: I also removed the static_argnums in `integtrators.py` which sets `step_size` (a float) as static. This was also causing a "non-hashable type" error for some reason. 
In any case, I don't think jitting floats adds a performance penalty (from playing around with Jax generally; I didn't test performance in this particular case).