@rlouf 2020-11-27
Since I put an emphasis on performance I should have benchmarks that are run with the CI at some point.