@jeremiecoullon 2020-11-27
I did some quick performance tests (on the example in the reamde):

• the "non-hashable type" and the `kernel_factory` fixes made no difference in performance (running on the old version Jax)
• the newer version of Jax was faster for both the compiled and non-compiled example.

### Details of test:

I ran 1D logistic regression (like in the readme) but with 2 chains and 3K samples:

**Old jax version:**
Non-compiled:
• warmup:  550 samples/sec
• sampling: 3800 samples/sec

Compiled:
• warmup: 1.9sec
• sampling: 1.7 sec

**New jax verson (0.2.6):**
Non-compiled:
• warmup: 720 samples/sec
• sampling: 5950 samples/sec

Complied:
• warmup:  1.5sec
• sampling: 1.3 sec

I ran these timings a few times and their variance was small; so the newer jax library definitely sped the sampler up in this example.
I don't know if it will make a difference for more complicated examples though!